+++ cmd/go-magnetar/main.go
@@ -8,6 +8,7 @@ import (
 	"github.com/wmentor/go-magnetar/internal/cmd"
 	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/compact"
 	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/exit"
+	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/fetch"
 	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/help"
 	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/less"
 	_ "github.com/wmentor/go-magnetar/internal/plugins/chatcmd/new"